Wed, Jun 22, 2022, 8:37 AM +00:00

+++ pset2/readability/readability.c
@@ -18,9 +18,9 @@ int main(void)
     int sentences = count_sentences(text);
 
     // Average number of letters per 100 words
-    int L = letters / words * 100
+    float L = letters / words * 100
     // Average number of sentences per 100 words
-    int S = sentences / words * 100
+    float S = sentences / words * 100
     //
     int index = 0.0588 * L - 0.296 * S - 15.8
